backend/src/events/api/serializers.py
from rest_framework import serializers

from events.models import Event


class EventSerializer(serializers.ModelSerializer):
    class Meta:
        model = Event
        fields = ('id', 'title', 'start_date',
                  'end_date', 'location', 'user_token')
